+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @lil-el/markdown
 
+## 1.3.0
+
+### Minor Changes
+
+- peer dependencies change, fix markdown height.
+
+### Patch Changes
+
+- Updated dependencies
+  - @lil-el/codepen@1.3.0
+
 ## 1.2.1
 
 ### Patch Changes